moving loadEnv outside
